refactor: chunk gen misc improvements
