+++ src/config/WorldConfig.ts
@@ -1,11 +1,6 @@
-
-
 export class WorldConfig {
-    static patchSize = Math.pow(2, 6)
-    static chunksYRange = {
-        min: 0,
-        max: 5
-    }
-    // max cache radius as a power of two
-    static cachePowLimit = 2// 4 => 16 patches radius
+  static patchSize = Math.pow(2, 6)
+
+  // max cache radius as a power of two
+  static cachePowLimit = 2 // 4 => 16 patches radius
 }
