+++ src/data/BoardContainer.ts
@@ -1,6 +1,6 @@
 import { Vector3 } from 'three'
 
-import { vect3ToVect2 } from '../common/utils'
+import { asVect2 } from '../common/utils'
 import { BlockType } from '../index'
 
 import { PatchContainer } from './DataContainers'
@@ -12,7 +12,7 @@ export class BoardContainer extends PatchContainer {
   constructor(center: Vector3, radius: number) {
     super()
     this.boardRadius = radius
-    this.boardCenter = vect3ToVect2(center).floor()
+    this.boardCenter = asVect2(center).floor()
     const board_dims = new Vector3(radius, 0, radius).multiplyScalar(2)
     this.bbox.setFromCenterAndSize(center.clone().floor(), board_dims)
     this.initFromBoxAndMask(this.bbox)
@@ -26,7 +26,7 @@ export class BoardContainer extends PatchContainer {
       const blocks = patch.iterOverBlocks(this.bbox)
       for (const block of blocks) {
         // discard blocs not included in board shape
-        const dist = vect3ToVect2(block.pos).distanceTo(boardCenter)
+        const dist = asVect2(block.pos).distanceTo(boardCenter)
         if (dist <= boardRadius) {
           const block_level = block.pos.y
           ymin = Math.min(block_level, ymin)
@@ -45,7 +45,7 @@ export class BoardContainer extends PatchContainer {
       const blocks = patch.iterOverBlocks(this.bbox)
       for (const block of blocks) {
         // discard blocs not included in board shape
-        const dist = vect3ToVect2(block.pos).distanceTo(boardCenter)
+        const dist = asVect2(block.pos).distanceTo(boardCenter)
         const y_diff = Math.abs(block.pos.y - avg)
         if (dist <= boardRadius && y_diff <= 5 && block.index !== undefined) {
           patch.writeBlockAtIndex(block.index, block.pos.y, BlockType.MUD)
