+++ src/index.ts
@@ -1,4 +1,3 @@
-
 export { Heightmap } from './procgen/Heightmap'
 export { NoiseSampler } from './procgen/NoiseSampler'
 export { ProcLayer } from './procgen/ProcLayer'
@@ -13,11 +12,11 @@ export { EntitiesMap, RepeatableEntitiesMap } from './procgen/EntitiesMap'
 export { EntityType } from './common/types'
 export { CacheContainer as WorldCacheContainer } from './data/CacheContainer'
 export { WorldComputeApi } from './compute/WorldComputeApi'
+export { ChunkFactory } from './tools/ChunkFactory'
+export { WorldConfig } from './config/WorldConfig'
 export * as WorldCompute from './compute/world-compute'
 export * as WorldUtils from './common/utils'
-export * as ChunkFactory from './tools/chunk-factory'
 export * as PlateauLegacy from './utils/plateau_legacy'
-export { WorldConfig } from './config/WorldConfig'
 
 // export type { MappingConf, MappingData, MappingRanges } from "./common/types"
 // export { DevHacks } from './tools/DevHacks'
